convert options as an array to pass in
I need to convert the default atom [] to be a more composable set of
options.  I don't know exactly how this will work, but I expect that I'd
provide a map of default options, and then be able to run through the
passed in args and project settings and mush them together.

Then, finally, I'd take that default set of options and convert that to
an array of options to the WsImport class.  Neat stuff this Clojure
language.