docs, default options, etc.
Add in a bunch of other options.

I also included docs for options I'm not going to parse right now, and
when I think about it, the :extra-options seems like it would be a good
catch-all in case they want to specify something or a bunch of
somethings for something I'm not mapping.