I'll use a transient until I figure out how not to…
I don't like how it works, but I would like to get around having to use
transient lists to get what I want out of it.  I'll have to ping
someone or dig out a better way of doing this.  It'll probably result
in some meta-abstraction whereupon I call out to something that recurs
and builds it up, but thinking on that seems kind of ugly.  Maybe
someone smarter in Clojure will tell me a better way.