flesh out more details on passing multiple WSDL files in
I'm still toying with this, but I think to start I'm going to assume
the same set of default options and then just pass in an array of WSDL
URL's/Filenames to parse out.  This way I can get to work on getting it
to generate something.

The second phase would be to find the best way to make a default set of
options wrap a user's default set of options, then to let the user
define their own options on a per-wsdl basis.  This would be a rewrite
of some functionality, however, so the goal at this point is to get as
thin as possible on function declarations so I don't have too many
places to switch an array listing with an array+wsdl option listing.