add failing tests for wsimport args
Add tests to parse out the wsimport task from the command-line.
Command line options are tricky, because the defaults probably won't be
always overridable (such as no explicit -NOT-Xnocompile exists), but
add tests nonetheless to see if it changes the defaults and adds the
appropriate extra options and wsdl properly.

Should add a test that fails if no .wsdl is supplied, but I don't know
how to do that yet in Midje.