doesn't feel so ugly to doseq
doseq seems appropriate in this instance, because it allows me to bind
a list of options through and call the same method on them.  I like
this, and it gives me thoughts about how I could possibly reorganize
this entire thing without using transients.  I'm still not sure what
that fix is yet, but that looks better.