add Amazon EC2 WSDL for testing
I need a few testing WSDL files just to get the generator for wsimport
to work, and this one does (albeit I have to pass in -extension to get
it to compile, it does it without blowing up).  I'm hoping this doesn't
violate some issue with open source or whatnot, but since I'm only
using it for testing and it's a publicly available web service, I don't
think it's a secret or some sort of NDA'd quality SOAP service that I'd
get a takedown notice.