add in jaxb-binding-files option
Leveraging my newfound knowledge of `doseq`, I can sort of get this to
do what I want to do, although I think there's probably a cleaner way
to do it.  I have to have an array ([]) passed in, however, in order
for this to work, so I had to change the tests (and unless I think it's
important, I'm not sure the user should assume a straight string option
OR an array of string options.  It's probably simpler to just assume
array).